Architektur umstrukturiert für Konsistenz der Aggregate

+++ src/main/java/de/propra/chicken/domain/model/KlausurRef.java
@@ -1,4 +1,4 @@
 package de.propra.chicken.domain.model;
 
-public record KlausurRef(Long id) {
+public record KlausurRef(int id) {
 }
